+++ docs/release-notes/release-notes.md
@@ -3,6 +3,10 @@
 This section lists the release notes for tagged version of **nvf** and the
 current main current main branch
 
+[JamyGolden](https://github.com/JamyGolden):
+
+- Add support for `tinted-vim` (base16, base24) themes
+
 ```{=include=} chapters
 rl-0.1.md
 rl-0.2.md